+++ module-app/src/main/java/maple/expectation/adapter/in/DlqPortAdapter.java
@@ -2,7 +2,7 @@ package maple.expectation.adapter.in;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import maple.expectation.controller.dto.common.CursorPageRequest;
+import maple.expectation.controller.dto.page.CursorPageRequest;
 import maple.expectation.core.port.inbound.DlqPort;
 import maple.expectation.service.v2.donation.outbox.DlqAdminService;
 import org.springframework.stereotype.Component;
