+++ module-app/src/main/java/maple/expectation/controller/dto/page/CursorPageResponse.java
@@ -1,0 +1,1 @@
+package maple.expectation.controller.dto.page;
