+++ module-app/src/main/java/maple/expectation/service/v2/donation/outbox/DlqAdminService.java
@@ -2,11 +2,11 @@ package maple.expectation.service.v2.donation.outbox;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import maple.expectation.controller.dto.common.CursorPageRequest;
-import maple.expectation.controller.dto.common.CursorPageResponse;
 import maple.expectation.controller.dto.dlq.DlqDetailResponse;
 import maple.expectation.controller.dto.dlq.DlqEntryResponse;
 import maple.expectation.controller.dto.dlq.DlqReprocessResult;
+import maple.expectation.controller.dto.page.CursorPageRequest;
+import maple.expectation.controller.dto.page.CursorPageResponse;
 import maple.expectation.domain.v2.DonationDlq;
 import maple.expectation.domain.v2.DonationOutbox;
 import maple.expectation.error.exception.DlqNotFoundException;
